# docs/swagger.yaml
basePath: /v1
definitions:
...
  internal_api.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  internal_api.GraphDataPoint:
    properties:
      avgResponseTime:
        description: Average response time in seconds
        type: number
      errorsPerSec:
        description: Errors per second
        type: number
      requestsPerSec:
        description: RPS
        type: number
      timestamp:
        description: Unix milliseconds
        type: integer
      users:
        description: Current active users
        type: integer
    type: object
...
  internal_api.RequestLogEntry:
    properties:
      requestType:
        description: GET, POST, etc.
        type: string
      responseLength:
        type: integer
      responseTime:
        description: In milliseconds
        type: number
      success:
        type: boolean
      timestamp:
        description: Unix milliseconds
        type: integer
      url:
        type: string
    type: object
  internal_api.RunGraphResponse:
    properties:
      dataPoints:
        items:
          $ref: '#/definitions/internal_api.GraphDataPoint'
        type: array
      runId:
        type: string
      runName:
        type: string
      startedAt:
        type: string
      status:
        type: string
    type: object
  internal_api.RunSummaryResponse:
    properties:
      avgResponseTime:
        description: In seconds
        type: number
      duration:
        type: string
      durationSeconds:
        type: integer
      errorRate:
        description: Percentage
        type: number
      finishedAt:
        type: string
      requestsPerSec:
        type: number
      runId:
        type: string
      runName:
        type: string
      spawnRate:
        type: number
      startedAt:
        type: string
      status:
        type: string
      targetUsers:
        description: Test configuration
        type: integer
      totalRequests:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@loadmanager.io
    name: API Support
  description: |-
    Load testing management platform for creating, executing, and monitoring distributed load tests with Locust.

    ## Features
    - **Load Test Management**: Create and manage load tests with script versioning
    - **Script Revisions**: Full version control for test scripts with audit trail
    - **Test Execution**: Start and monitor load test runs with real-time metrics
    - **Visualization**: Real-time dashboards and historical metrics

    ## Authentication
    Currently, this API does not require authentication. Future versions will include API key or OAuth2 support.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Load Manager API
  version: "1.0"
paths:
...
  /runs/{id}/graph:
    get:
      description: Returns minimal graph data optimized for dashboard charts (RPS
        and response time over time)
      parameters:
      - description: Load Test Run ID
        in: path
        name: id
        required: true
        type: string
      - description: Start time in RFC3339 format
        in: query
        name: from
        type: string
      - description: End time in RFC3339 format
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Graph data for visualization
          schema:
            $ref: '#/definitions/internal_api.RunGraphResponse'
        "404":
          description: Load test run not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Failed to fetch graph data
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get run graph data
      tags:
      - Visualization
...
  /runs/{id}/requests:
    get:
      description: Returns endpoint statistics in a log-like format showing performance
        per endpoint
      parameters:
      - description: Load Test Run ID
        in: path
        name: id
        required: true
        type: string
      - default: 100
        description: Maximum number of entries to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of request statistics per endpoint
          schema:
            items:
              $ref: '#/definitions/internal_api.RequestLogEntry'
            type: array
        "404":
          description: Load test run not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Failed to fetch request log
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get live request statistics
      tags:
      - Visualization
...
  /runs/{id}/summary:
    get:
      description: Returns the 4 key metrics for dashboard cards (total requests,
        RPS, avg response time, error rate)
      parameters:
      - description: Load Test Run ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Summary metrics
          schema:
            $ref: '#/definitions/internal_api.RunSummaryResponse'
        "404":
          description: Load test run not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Failed to fetch summary
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get run summary metrics
      tags:
      - Visualization
schemes:
- http
- https
swagger: "2.0"
...